refactor(lib): extract self-update + auto-update-check into self_update (#217)
Move run_self_update (the explicit `rvpm self-update` command) plus the
background auto-update probe (AutoUpdateHandle, auto_update_disabled_by_env,
maybe_spawn_auto_update_check, config_path_for_auto_check,
finalize_auto_update_check) out of lib.rs into src/self_update.rs.

All pub(crate), glob-imported by lib.rs. paths/config helper calls are
qualified to crate::paths:: / crate::config::.

Behaviour-preserving code motion. fmt --check + clippy -D warnings + test (805)
+ --doc all green.
